fix: Fixed a problem where `.gitignore` was not included in the created project
