+++ src/thematic-markdown.ts
@@ -1,5 +1,5 @@
 import upath from 'upath'
-import fs from 'fs'
+import fs from 'node:fs'
 
 /** Config filename of Vivliostyle. */
 const CONFIG_FILE_NAME = 'vivliostyle.config.js'
@@ -18,7 +18,7 @@ type Entry = string | { path?: string }
 
 /**
  * Get the path to the directory of the theme installed in the project generated by Create Book.
- * @param projectDir The path to the project directory generated by Create Book.
+ * @param projectDir - The path to the project directory generated by Create Book.
  * @returns A path string on success, an empty string on failure.
  * @throws Failed to load `vivliostyle.config.js`.
  */
@@ -33,7 +33,7 @@ const getThemeDir = (projectDir: string): string => {
 
 /**
  * Rename the default Markdown so it doesn't conflict with the theme's.
- * @param projectDir The path to the project directory generated by Create Book.
+ * @param projectDir - The path to the project directory generated by Create Book.
  */
 const escapeDefaultMarkdownFile = (projectDir: string) => {
   const src = upath.join(projectDir, DEFAULT_MARKDOWN_FILE_NAME)
@@ -43,8 +43,8 @@ const escapeDefaultMarkdownFile = (projectDir: string) => {
 
 /**
  * Restore the Markdown file in your project's directory to its original state.
- * @param projectDir The path to the project directory generated by Create Book.
- * @param entries Markdown file name list copied from theme to project directory.
+ * @param projectDir - The path to the project directory generated by Create Book.
+ * @param entries - Markdown file name list copied from theme to project directory.
  */
 const restoreDefaultMarkdownFile = (projectDir: string, entries: Entry[]) => {
   entries.forEach((entry) => {
@@ -62,8 +62,8 @@ const restoreDefaultMarkdownFile = (projectDir: string, entries: Entry[]) => {
 
 /**
  * Copy the Markdown files.
- * @param projectDir The path to the project directory generated by Create Book.
- * @param entries Relative path of the Markdown files on theme directory, e.g. '["example/foo.md", "bar.md"]'
+ * @param projectDir - The path to the project directory generated by Create Book.
+ * @param entries - Relative path of the Markdown files on theme directory, e.g. '["example/foo.md", "bar.md"]'
  * @returns Name of the copied files.
  */
 const copyMarkdownFiles = (
@@ -108,8 +108,8 @@ const copyMarkdownFiles = (
 
 /**
  * Replace `entry` in your project's config file with a copied Markdown file.
- * @param projectDir The path to the project directory generated by Create Book.
- * @param entries Name of the Markdown files. e.g. '["foo.md", "bar.md"]'
+ * @param projectDir - The path to the project directory generated by Create Book.
+ * @param entries - Name of the Markdown files. e.g. '["foo.md", "bar.md"]'
  */
 const replaceConfigEntries = (projectDir: string, entries: Entry[]) => {
   const filePath = upath.join(projectDir, CONFIG_FILE_NAME)
@@ -137,7 +137,7 @@ const replaceConfigEntries = (projectDir: string, entries: Entry[]) => {
 /**
  * Apply thematic Markdown.
  * The markdown file is based on the `vivliostyle.config.js` definition of the installed theme.
- * @param projectDir The path to the project directory generated by Create Book.
+ * @param projectDir - The path to the project directory generated by Create Book.
  * @returns `true` on success, `false` on failure.
  * @throws Failed to load `vivliostyle.config.js`.
  */
