+++ templates/default/gitignore
